% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data_wroc.R
\docType{data}
\name{example_data_wroc}
\alias{example_data_wroc}
\title{Simulated data}
\format{
\subsection{\code{example_data_wroc}}{

A data frame with 740 rows and 3 columns:
\describe{
\item{y}{Response variable}
\item{phat}{Predicted probabilities}
\item{weights}{Sampling weights}
...
}
}
}
\usage{
example_data_wroc
}
\description{
This dataset has been simulated in order to provide the users with an example dataset.
}
\keyword{datasets}
