% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_variables_wroc_2.R
\docType{data}
\name{example_variables_wroc_2}
\alias{example_variables_wroc_2}
\title{Simulated data}
\format{
\subsection{\code{example_variables_wroc_2}}{

A data frame with 1720 rows and 10 columns:
\describe{
\item{y}{Response variable}
\item{x1,...,x6}{Covariates}
\item{strata}{Strata variable}
\item{cluster}{Cluster variable}
\item{weights}{Sampling weights}
...
}
}
}
\usage{
example_variables_wroc_2
}
\description{
This dataset has been simulated in order to provide the users with an example dataset.
}
\keyword{datasets}
