% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyroc.R
\name{svyroc}
\alias{svyroc}
\title{Estimate a survey-weighted ROC curve}
\usage{
svyroc(
  design,
  prediction,
  label,
  fpr_grid = NULL,
  se_method = NULL,
  n_boot = 250
)
}
\arguments{
\item{design}{survey.design object specifying the complex survey design}

\item{prediction}{character value specifying the name of the variable containing the predicted value}

\item{label}{character value specifying the name of the binary response encoded 0 (control) and 1 (cases)}

\item{fpr_grid}{NULL (default) or numeric vector ranging from 0 to 1 specifying the desired cutpoints for the FPR.
If set to NULL, the cutpoints are defined by the unique values of the predictor}

\item{se_method}{NULL (default) if standard error is not necessary; "asymptotic" to compute asymptotic SE, and "bootstrap" for bootstrap SE}

\item{n_boot}{number of bootstrap replicates used to compute SE when se_method = "bootstrap"}
}
\value{
list containing the ROC curve and the corresponding AUC
}
\description{
Constructs the ROC curve to a data from complex survey design with asymptotic design-based standard error
}
