% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyroc_auc.R
\name{svyroc_auc}
\alias{svyroc_auc}
\title{Function to compute Area Under the Curve using trapezoid method}
\usage{
svyroc_auc(fpr, tpr)
}
\arguments{
\item{fpr}{a numeric vector with false positive rate (FPR) from the ROC curve}

\item{tpr}{a numeric vector with true positive rate (TPR) from the ROC curve

A function that computes the numeric value under the ROC curve with trapezoidal rule}
}
\value{
a numeric value for the area under the ROC curve (AUC-ROC)
}
\description{
Function to compute Area Under the Curve using trapezoid method
}
