% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyroc_estimate.R
\name{svyroc_estimate}
\alias{svyroc_estimate}
\title{Estimate a survey-weighted ROC curve}
\usage{
svyroc_estimate(design, prediction, label, fpr_grid = NULL)
}
\arguments{
\item{design}{survey.design object specifying the complex survey design}

\item{prediction}{character value specifying the name of the variable containing the predicted value}

\item{label}{character value specifying the name of the binary response encoded 0 (control) and 1 (cases)}

\item{fpr_grid}{NULL (default) or numeric vector ranging from 0 to 1 specifying the desired cutpoints for the FPR.
If set to NULL, the cutpoints are defined by the unique values of the predictor

This function estimates the ROC curve non-parametrically while accounting for the complex sampling design}
}
\value{
A list containing the ROC curve and the corresponding AUC
}
\description{
Estimate a survey-weighted ROC curve
}
