% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyrocvar.R
\name{svyroc_var}
\alias{svyroc_var}
\title{Function to compute design-adjusted variance for ROC curve}
\usage{
svyroc_var(roc_data, design, fpr_grid)
}
\arguments{
\item{roc_data}{output from svyroc_estimate}

\item{design}{survey.design object specifying the complex survey design}

\item{fpr_grid}{NULL (default) or numeric vector ranging from 0 to 1 specifying the desired cutpoints for the FPR.
If set to NULL, the cutpoints are defined by the unique values of the predictor}
}
\description{
Function to compute design-adjusted variance for ROC curve
}
